+++ CHANGELOG.md
@@ -5,6 +5,10 @@ documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- Continue [migration from Travis-CI to Github-Actions][Run-Travis-Yml]
+
 ## [0.0.51] - 2020-12-09
 ### Added
 - Support for PHP 8