+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Continue [migration from Travis-CI to Github-Actions][Run-Travis-Yml]
 - Rename tests folder to test to streamline directory names.
+### Fixed
+- Phpunit test-case shim for invalid-argument-helper since Phpunit
+  6.x, missing in 0.0.51
 
 ## [0.0.51] - 2020-12-09
 ### Added